Changed the two calls to "fabsf" to "fabs". I really see no reason why "fabsf" should be used, especially since on at least one Linux system (lightning, x86_64 bit), this function reports the wrong results if you compile with "-ansi".   For example, "fabsf(-100.)" returns "1" if you compile with "-ansi", whereas "fabs(-100.)" returns "100" either way.
This is inside the wrapper for the "abs" built-in function.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9122 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/BuiltInFuncs.c
@@ -1,5 +1,5 @@
 /*
- *      $Id: BuiltInFuncs.c,v 1.205 2006-08-31 23:50:30 dbrown Exp $
+ *      $Id: BuiltInFuncs.c,v 1.206 2006-09-03 20:03:03 haley Exp $
  */
 /************************************************************************
 *                                                                       *
@@ -5011,14 +5011,14 @@ NhlErrorTypes _NclIabs
             if (has_missing) {
                 for (i = 0; i < total; i++) {
                     if (fvalue[i] != missing.floatval) {
-                        fout_val[i] = (float) fabsf((float) fvalue[i]);
+                        fout_val[i] = (float) fabs((float) fvalue[i]);
                     } else {
                         fout_val[i] = (float) missing.floatval;
                     }
                 }
             } else {
                 for (i = 0; i < total; i++) {
-                    fout_val[i] = (float) fabsf((float) fvalue[i]);
+                    fout_val[i] = (float) fabs((float) fvalue[i]);
                 }
             }
